Remove nested function declaration (#7)
* Remove nested function declaration

I was playing around with the npm5 and a few cases where it should be blazingly fast are still a bit sluggish. This package is used many many times in tight loops in npm, especially module lookups. Avoiding nested function declarations speeds this module up a fair amount. Let me know if I missed something or if you need some benchs to prove it. Thanks for considering this patch and all the awesome work you do!